+++ backend/dataall/modules/datapipelines/api/resolvers.py
@@ -1,8 +1,8 @@
 import logging
 
 from dataall.base.api.context import Context
-from dataall.core.stacks.api import stack_helper
 from dataall.base.db import exceptions
+from dataall.core.stacks.services.stack_service import StackService
 from dataall.modules.datapipelines.api.enums import DataPipelineRole
 from dataall.modules.datapipelines.db.datapipelines_models import DataPipeline
 from dataall.modules.datapipelines.services.datapipelines_service import DataPipelineService
@@ -102,7 +102,7 @@ def resolve_clone_url_http(context: Context, source: DataPipeline, **kwargs):
 def resolve_stack(context, source: DataPipeline, **kwargs):
     if not source:
         return None
-    return stack_helper.get_stack_with_cfn_resources(
+    return StackService.get_stack_with_cfn_resources(
         targetUri=source.DataPipelineUri,
         environmentUri=source.environmentUri,
     )
